Add display group test utilities
